backend/requirements.txt
# Flask backend dependencies
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
pymongo==4.7.2
bcrypt==4.1.3
python-dotenv==1.0.1
requests==2.31.0
Werkzeug==3.0.3
